+++ sermil-core/src/br/mil/eb/sermil/modelo/CidEvento.java
@@ -32,11 +32,13 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
 
    private static final long serialVersionUID = -9003467429456295200L;
 
+   /* Deprecated: usar Enum TipoEvento
    public final static Byte ALISTAMENTO = 1;
    public final static Byte DISPENSA_SELECAO = 3; 
    public final static Byte EXCESSO_CONTINGENTE = 6;
    public final static Byte LICENCIAMENTO = 12; 
-
+   */
+   
    @EmbeddedId
    private CidEvento.PK pk;
 
@@ -53,7 +55,7 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
       this.setPk(new CidEvento.PK());
    }
 
-   public CidEvento(final Long ra, final Byte codigo, final Date data) {
+   public CidEvento(final Long ra, final Integer codigo, final Date data) {
       this.setPk(new CidEvento.PK(ra, codigo, data));
    }
 
@@ -144,7 +146,7 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
       private Long cidadaoRa;
 
       @Column(name = "CODIGO")
-      private Byte codigo;
+      private Integer codigo;
 
       @Temporal(TemporalType.DATE)
       @Column(name = "DATA")
@@ -154,7 +156,7 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
          super();
       }
 
-      public PK(final Long cidadaoRa, final Byte codigo, final Date data) {
+      public PK(final Long cidadaoRa, final Integer codigo, final Date data) {
          super();
          this.setCidadaoRa(cidadaoRa);
          this.setCodigo(codigo);
@@ -214,7 +216,7 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
          return this.cidadaoRa;
       }
 
-      public Byte getCodigo() {
+      public Integer getCodigo() {
          return this.codigo;
       }
 
@@ -226,7 +228,7 @@ public final class CidEvento implements Comparable<CidEvento>, Serializable {
          this.cidadaoRa = cidadaoRa;
       }
 
-      public void setCodigo(Byte codigo) {
+      public void setCodigo(Integer codigo) {
          this.codigo = codigo;
       }
